+++ src/Toolkit/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.0.0
 
 -   Minimum required Symfony version is now 6.4
+-   Minimum required PHP version is now 8.2
 
 # 2.29.0